# api/handlers/v1/profession.go
package handlers

import (
	"api-gateway/api/models"
	"api-gateway/genproto/position_service"
	"api-gateway/pkg/util"
	"context"
	"errors"
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
)

// CreateProfession godoc
// @ID create-profession
// @Router /v1/profession [POST]
// @Summary create profession
// @Description Create Profession
// @Tags profession
// @Accept json
// @Produce json
// @Param profession body models.CreateProfession true "profession"
// @Success 200 {object} models.ResponseModel{data=string} "desc"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Failure 500 {object} models.ResponseModel{error=string} "Server Error"
func (h *handler) CreateProfession(c *gin.Context) {
	var position models.CreateProfession
	err := c.BindJSON(&position)
	if err != nil {
		h.handleErrorResponse(c, 400, "error wrong input ", err)
		return
	}

	resp, err := h.services.ProfessionService().Create(
		context.Background(),
		&position_service.CreateProfession{
			Name: position.Name,
		},
	)
	if !handleError(h.log, c, err, "error while creating profession") {
		return
	}

	h.handleSuccessResponse(c, http.StatusOK, "ok", resp)
}

// GetAllProfession godoc
// @ID get-all-profession
// @Router /v1/profession [GET]
// @Summary get all profession
// @Description Get All Profession
// @Tags profession
// @Accept json
// @Produce json
// @Param name query string false "name"
// @Param limit query string false "limit"
// @Param offset query string false "offset"
// @Success 200 {object} models.ResponseModel{data=models.GetAllProfessionResponse} "desc"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Failure 500 {object} models.ResponseModel{error=string} "Server Error"
func (h *handler) GetAllProfession(c *gin.Context) {
	var professions models.GetAllProfessionResponse

	limit, err := h.ParseQueryParam(c, "limit", "10")
	if err != nil {
		return
	}

	offset, err := h.ParseQueryParam(c, "offset", "0")
	if err != nil {
		return
	}
	resp, err := h.services.ProfessionService().GetAll(
		context.Background(),
		&position_service.GetAllProfessionRequest{
			Limit:  int64(limit),
			Offset: int64(offset),
			Name:   c.Query("name"),
		},
	)
	if !handleError(h.log, c, err, "error while getting all professions") {
		fmt.Println("connection error")
		return
	}

	err = ParseToStruct(&professions, resp)
	if err != nil {
		h.handleErrorResponse(c, http.StatusInternalServerError, "error while parsing to struct", err)
		return
	}

	h.handleSuccessResponse(c, http.StatusOK, "ok", professions)

}

// GetProfession godoc
// @ID get-profession
// @Router /v1/profession/{profession_id} [GET]
// @Summary get profession by id
// @Description Get Profession By Id
// @Tags profession
// @Accept json
// @Produce json
// @Param profession_id path string true "profession_id"
// @Success 200 {object} models.ResponseModel{data=models.Profession} "desc"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Failure 500 {object} models.ResponseModel{error=string} "Server Error"
func (h *handler) GetProfession(c *gin.Context) {
	var profession models.Profession
	idStr := c.Param("profession_id")

	if !util.IsValidUUID(idStr) {
		h.handleErrorResponse(c, http.StatusBadRequest, "profession_id is not valid ", errors.New("profession_id is not valid"))
		return
	}

	resp, err := h.services.ProfessionService().GetById(
		context.Background(),
		&position_service.ProfessionId{
			Id: idStr,
		},
	)

	if !handleError(h.log, c, err, "error from getting profession") {
		return
	}

	err = ParseToStruct(&profession, resp)
	if err != nil {
		h.handleErrorResponse(c, http.StatusInternalServerError, "error parsing ", err)
		return
	}
	profession.Name = ""
	h.handleSuccessResponse(c, http.StatusOK, "ok", profession)

}

// UpdateProfession godoc
// @ID update-profession
// @Router /v1/profession [PUT]
// @Summary update profession
// @Description Update Profession
// @Tags profession
// @Accept json
// @Produce json
// @Param profession body models.Profession true "profession"
// @Success 200 {object} models.ResponseModel{data=models.MsgModel} "desc"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Failure 500 {object} models.ResponseModel{error=string} "Server Error"
func (h *handler) UpdateProfession(c *gin.Context) {
	var profession models.Profession
	if err := c.BindJSON(&profession); err != nil {
		h.handleErrorResponse(c, 400, "wrong input for profession", err)
		return
	}

	_, err := h.services.ProfessionService().Update(
		context.Background(),
		&position_service.Profession{
			Id:   profession.Id,
			Name: profession.Name,
		},
	)
	if !handleError(h.log, c, err, "error Update profession") {
		return
	}
	h.handleSuccessResponse(c, 200, "Updated", models.MsgModel{Msg: "Updated"})
}

// DeleteProfession godoc
// @ID delete-profession
// @Router /v1/profession/{profession_id} [DELETE]
// @Summary delete profession
// @Description Delete Profession
// @Tags profession
// @Accept json
// @Produce json
// @Param profession_id path string true "profession_id"
// @Success 200 {object} models.ResponseModel{data=models.MsgModel} "desc"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Response 400 {object} models.ResponseModel{error=string} "Bad Request"
// @Failure 500 {object} models.ResponseModel{error=string} "Server Error"
func (h *handler) DeleteProfession(c *gin.Context) {
	profession_id := c.Param("profession_id")
	if !util.IsValidUUID(profession_id) {
		h.handleErrorResponse(c, http.StatusBadRequest, "profession_id is not valid ", errors.New("profession_id is not valid"))
		return
	}

	_, err := h.services.ProfessionService().Delete(
		context.Background(),
		&position_service.ProfessionId{
			Id: profession_id,
		},
	)

	fmt.Println("after resp")
	if !handleError(h.log, c, err, "error while deleting profession") {
		fmt.Println("error 1")
		return
	}

	fmt.Println("before return ")
	h.handleSuccessResponse(c, 200, "deleted", models.MsgModel{Msg: "Deleled"})

}